+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: pavorapa <pavorapa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/19 18:31:09 by pavorapa          #+#    #+#             */
-/*   Updated: 2024/12/03 21:32:40 by pavorapa         ###   ########.fr       */
+/*   Updated: 2024/12/03 23:12:49 by pavorapa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (c[i])
